requirements.txt
boto3==1.33.9
celery==5.3.5
Django==4.2.4
django-admin-rangefilter==0.12.0
django-cors-headers==4.3.1
django-jazzmin==2.6.0
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
psycopg2-binary==2.9.9
Pillow==10.1.0
python-decouple==3.8
redis==5.0.1